+++ org.eclipse.buildship.core/src/main/java/org/eclipse/buildship/core/internal/workspace/ModelProviderUtil.java
@@ -9,7 +9,6 @@
 package org.eclipse.buildship.core.internal.workspace;
 
 import java.util.Collection;
-import java.util.Set;
 
 import org.gradle.tooling.BuildAction;
 import org.gradle.tooling.CancellationTokenSource;
@@ -18,19 +17,17 @@ import org.gradle.tooling.model.build.BuildEnvironment;
 import org.gradle.tooling.model.eclipse.EclipseProject;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 
 import org.eclipse.buildship.core.internal.util.gradle.BuildActionUtil;
 import org.eclipse.buildship.core.internal.util.gradle.GradleVersion;
-import org.eclipse.buildship.core.internal.util.gradle.HierarchicalElementUtils;
 import org.eclipse.buildship.core.internal.util.gradle.SimpleIntermediateResultHandler;
 
 public class ModelProviderUtil {
 
-    public static Set<EclipseProject> fetchAllEclipseProjectsWithSyncTask(InternalGradleBuild build, CancellationTokenSource tokenSource, FetchStrategy fetchStrategy, IProgressMonitor monitor) throws Exception {
-        Collection<EclipseProject> models = build.withConnection(connection -> {
+    public static Collection<EclipseProject> fetchAllEclipseProjectsWithSyncTask(InternalGradleBuild build, CancellationTokenSource tokenSource, FetchStrategy fetchStrategy, IProgressMonitor monitor) throws Exception {
+        return build.withConnection(connection -> {
             BuildEnvironment buildEnvironment = connection.getModel(BuildEnvironment.class);
             GradleVersion gradleVersion = GradleVersion.version(buildEnvironment.getGradle().getGradleVersion());
 
@@ -42,7 +39,7 @@ public class ModelProviderUtil {
                 return queryEclipseModel(connection);
             }
         }, monitor);
-        return collectAll(models);
+
     }
 
     private static boolean supportsTaskExecution(GradleVersion gradleVersion) {
@@ -68,13 +65,5 @@ public class ModelProviderUtil {
 
     private static Collection<EclipseProject> queryEclipseModel(ProjectConnection connection) {
         return ImmutableList.of(connection.getModel(EclipseProject.class));
-    }
-
-    private static Set<EclipseProject> collectAll(Collection<EclipseProject> models) {
-        ImmutableSet.Builder<EclipseProject> result = ImmutableSet.builder();
-        for (EclipseProject model : models) {
-            result.addAll(HierarchicalElementUtils.getAll(model));
-        }
-        return result.build();
     }
 }
