+++ org.eclipse.buildship.ui.test/src/main/groovy/org/eclipse/buildship/ui/internal/view/task/BuildInvocationsTest.groovy
@@ -161,7 +161,7 @@ class BuildInvocationsTest extends ProjectSynchronizationSpecification {
     }
 
     private static Set<String> getImplicitTasks() {
-        ["init", "wrapper", "help", "projects", "tasks", "properties", "components", "dependencies", "dependencyInsight", "setupBuild", "model", "buildEnvironment", "dependentComponents"]
+        ["init", "wrapper", "help", "projects", "tasks", "properties", "components", "dependencies", "dependencyInsight", "setupBuild", "model", "buildEnvironment", "dependentComponents", "nothing", "prepareKotlinBuildScriptModel"]
     }
 
     private static Set<String> collectNamesOfNonImplicitProjectTasks(List<ProjectTask> tasks) {
